+++ packages/sdnext/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-node
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.47.2
+
 ## 0.0.15
 
 ### Patch Changes